Should resolve #35 clang-tidy warnings
- added majority of available clang-tidy checks
- disabled make_unique warnings as basalt uses C++11
- disabled google-runtime-references
- disabled clang-diagnostic-ignored-optimization-argument
- .clang-tidy list of checks now one item per line
- added SYSTEM to 3rd party CMake include_directories

Changed code for the following types of warnings:
- pass by const ref -> pass by copy & std::move
- if {return} else {return} -> if {return} return
- added some // NOLINTS(..) to reduce third party warnings
- global static const -> functions returning ref to static var
- raw owning pointers -> unique_ptrs
- using namespace X -> using X::Y
- added more // NOLINTs to supress catch2 warnings

Code now compiles without any clang-tidy warnings for:
- clang6
- clang7
- gcc7

+++ src/basalt/py_helpers.hpp
@@ -191,7 +191,7 @@ void deserialize_vector(std::istream& iss, std::vector<T>& data) {
 template <typename T>
 void fill_vector(pybind11::array_t<T>& array, std::vector<T>& vector) {
     auto buffer = array.request(false);
-    vector.reserve(buffer.size);
+    vector.reserve(static_cast<std::size_t>(buffer.size));
     auto ptr = reinterpret_cast<typename std::add_pointer<T>::type>(buffer.ptr);
     std::copy(ptr, ptr + buffer.size, std::back_inserter(vector));
 }
